// types/Vizity.ts
export interface ISettingFields {
    akciya: IAkciyaFields[] | null;
    oprosnik: IOprosnikFields[] | null;
    deteiling: IDeteilingFields[] | null;
}